+++ photutils/psf.py
@@ -597,9 +597,8 @@ def prepare_psf_model(psfmodel, xname=None, yname=None, fluxname=None,
 def psf_photometry(data, positions, psf, fitshape=None,
                    fitter=LevMarLSQFitter(),
                    unit=None, wcs=None, error=None, effective_gain=None,
-                   mask=None, pixelwise_error=True,
-                   mode='sequential',
-                   store_fit_info=False):
+                   mask=None, pixelwise_error=True, mode='sequential',
+                   store_fit_info=False, param_uncert=False):
     """
     Perform PSF/PRF photometry on the data.
 
@@ -686,6 +685,11 @@ def psf_photometry(data, positions, psf, fitshape=None,
         table will have an additional column 'fit_message' with the message that
         came from the fit.  If a list, it will be populated with the
         ``fit_info`` dictionary of the fitter for each fit.
+    param_uncert : bool (default=False)
+        If True, the uncertainties on each parameter estimate will be stored
+        in the output table. This option assumes that the fitter has the
+        'param_cov' key in its 'fit_info' dictionary.
+        See 'fit_info' in `~astropy.modeling.fitting.LevMarLSQFitter`.
 
     Returns
     -------
@@ -773,6 +777,15 @@ def psf_photometry(data, positions, psf, fitshape=None,
         fit_infos = store_fit_info
     elif store_fit_info:
         fit_messages = []
+    if param_uncert:
+        if 'param_cov' in fitter.fit_info:
+            uncert = []
+        else:
+            warnings.warn("uncertainties on fitted parameters cannot be " +
+                           "computed because fitter does not contain " +
+                           "`param_cov` key in its `fit_info` dictionary.",
+                           AstropyUserWarning)
+            param_uncert = False
 
     # Many fitters take a "weight" array, but no "mask".
     # Thus, we convert the mask to weights on 1 and 0. Unfortunately,
@@ -808,11 +821,28 @@ def psf_photometry(data, positions, psf, fitshape=None,
                 fit_infos.append(fitter.fit_info)
             if fit_messages is not None:
                 fit_messages.append(fitter.fit_info['message'])
+            if param_uncert:
+                if fitter.fit_info['param_cov'] is not None:
+                    uncert.append(np.sqrt(np.diag(\
+                            fitter.fit_info['param_cov'])))
+                else:
+                    warnings.warn("uncertainties on fitted parameters " +
+                                  "cannot be computed because the fit may " +
+                                  "be unsuccessful", AstropyUserWarning)
+                    uncert.append((None, None, None))
     else:
         raise ValueError('Invalid photometry mode.')
 
     if fit_messages is not None:
         result_tab['fit_messages'] = fit_messages
+    if param_uncert:
+        uncert = np.array(uncert)
+        result_tab.add_column(Column(name='flux_uncertainty', unit=fluxunit,
+                                     data=uncert[:, 0]))
+        result_tab.add_column(Column(name='x_uncertainty', unit=None,
+                                     data=uncert[:, 1]))
+        result_tab.add_column(Column(name='y_uncertainty', unit=None,
+                                     data=uncert[:, 2]))
 
     return result_tab